cifar10/requirements.txt
opencv-python
Pillow
matplotlib

scikit-learn
numpy
scipy
pandas

pyprind
gdown